pyproject.toml
[project]
name = "study-guide-agent"
version = "0.1.0"
description = "Provider-agnostic Quercus study guide agent"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
]

[project.optional-dependencies]
gcp = ["google-cloud-storage>=2.18.0"]
azure = [
  "azure-storage-blob>=12.23.0",
  "azure-identity>=1.17.0",
  "openai>=1.40.0",
]
dev = ["pytest>=8.3.0"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
